Fix - Header file

+++ 0x1E-search_algorithms/0-linear.c
@@ -9,14 +9,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (!array)
 		return (-1);
 
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%lu] = [%i]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
